update latlong primitives
