+++ featuretools/primitives/standard/transform_primitives/__init__.py
@@ -2,4 +2,5 @@
 from featuretools.primitives.standard.transform_primitives.binary_transform_primitives import *
 from featuretools.primitives.standard.transform_primitives.cumulative_transform_primitives import *
 from featuretools.primitives.standard.transform_primitives.exponential_transform_primitives import *
+from featuretools.primitives.standard.transform_primitives.latlong_transform_primitives import *
 from featuretools.primitives.standard.transform_primitives.natural_language_primitives import *
