+++ featuretools/primitives/standard/transform_primitives/latlong_transform_primitives/__init__.py
@@ -1,0 +1,18 @@
+from featuretools.primitives.standard.transform_primitives.latlong_transform_primitives.cityblock_distance import (
+    CityblockDistance,
+)
+from featuretools.primitives.standard.transform_primitives.latlong_transform_primitives.geomidpoint import (
+    GeoMidpoint,
+)
+from featuretools.primitives.standard.transform_primitives.latlong_transform_primitives.haversine import (
+    Haversine,
+)
+from featuretools.primitives.standard.transform_primitives.latlong_transform_primitives.is_in_geobox import (
+    IsInGeoBox,
+)
+from featuretools.primitives.standard.transform_primitives.latlong_transform_primitives.latitude import (
+    Latitude,
+)
+from featuretools.primitives.standard.transform_primitives.latlong_transform_primitives.longitude import (
+    Longitude,
+)
